pyproject.toml
[project]
name = "ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "google-genai==1.12.1",
    "pytest>=8.4.1",
    "python-dotenv==1.1.0",
    "requests>=2.32.4",
]
